+++ api/memory/manager/__init__.py
@@ -1,10 +1,10 @@
 """记忆管理器子包 — 抽象基类与 YAML 实现。"""
 
 from api.memory.manager.base import BaseMemoryManager
+from api.memory.manager.builder import MemoryManagerBuilder
 from api.memory.manager.yaml import (
     MAX_DESC_LENGTH,
     MemoryItem,
-    MemoryManagerBuilder,
     YamlMemoryManager,
 )
